Remove ApplicationComponentStreams
