+++ app/components/op_primer/banner_message_component.rb
@@ -26,16 +26,18 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-module ApplicationComponentStreams
-  extend ActiveSupport::Concern
+module OpPrimer
+  class BannerMessageComponent < Primer::Alpha::Banner
+    include ApplicationHelper
+    include OpTurbo::Streamable
+    include OpPrimer::ComponentHelpers
 
-  included do
-    def render_error_flash_message_via_turbo_stream(**kwargs)
-      update_flash_message_via_turbo_stream(**kwargs.merge(scheme: :danger, icon: :stop))
-    end
+    def initialize(**system_arguments)
+      system_arguments[:test_selector] ||= "primer-banner-message-component"
+      system_arguments[:dismiss_scheme] ||= :remove
+      system_arguments[:dismiss_label] ||= I18n.t(:button_close)
 
-    def update_flash_message_via_turbo_stream(**)
-      turbo_streams << turbo_stream.prepend("flash-container", Meetings::UpdateFlashComponent.new(**).render_in(view_context))
+      super
     end
   end
 end
