+++ app/controllers/admin/custom_fields/custom_field_projects_controller.rb
@@ -31,7 +31,6 @@
 class Admin::CustomFields::CustomFieldProjectsController < ApplicationController
   include OpTurbo::ComponentStream
   include OpTurbo::DialogStreamHelper
-  include ApplicationComponentStreams
   include FlashMessagesOutputSafetyHelper
 
   layout "admin"
