--- modules/meeting/app/components/meetings/update_flash_component.html.erb
@@ -1,19 +1,0 @@
-<div class="op-primer-banner--item">
-  <%=
-    render(Primer::Alpha::Banner.new(
-      full: false,
-      icon: :info,
-      scheme: :default,
-      data: {
-        singleton: true
-      }
-    )) do |component|
-      component.with_action_button(
-        tag: :a,
-        href: meeting_path(@meeting),
-        data: { turbo: false },
-        size: :medium) { I18n.t("label_meeting_reload") }
-      I18n.t("notice_meeting_updated")
-    end
-  %>
-</div>
