+++ modules/meeting/app/controllers/meeting_agenda_items_controller.rb
@@ -30,7 +30,6 @@ class MeetingAgendaItemsController < ApplicationController
   include AttachableServiceCall
   include OpTurbo::ComponentStream
   include Meetings::AgendaComponentStreams
-  include ApplicationComponentStreams
 
   before_action :set_meeting
   before_action :set_agenda_item_type, only: %i[new create]
